The core functionality is now there
It gets the average ages of the actors.  Unfortunately, it's slow as
molasses.  I will look into speeding it up some, but I think it may just
come down to the fact that I am pulling so many pages from IMDb.  I
could just rewrite that section to pull from a site with a master list,
or I could just maintain my own master list, but right now I'm just glad
it works.  The next step is to make it look pretty.